+++ src/Core/Controller/AbstractApplicationController.php
@@ -4,9 +4,21 @@ declare(strict_types=1);
 
 namespace App\Core\Controller;
 
+use App\Authentication\Entity\User;
 use CloudBase\LatteHelper\Controller\AbstractLatteController;
 
 class AbstractApplicationController extends AbstractLatteController
 {
     protected string $templateDir = '/templates';
+
+    protected function getUser(): ?User
+    {
+        $user = parent::getUser();
+
+        if (!$user instanceof User) {
+            return null;
+        }
+
+        return $user;
+    }
 }