+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added resend verification email button and functionality
+- Added a new `dtmVerified` column to the verification entity 
+- Implemented time limit between token requests
+
+### Changed
+- Renamed `VerificationToken` to `EmailVerificationToken` to be more explicit
+
 ## [2.1.0] - 2026-03-30
 ### Added
 - Added `Role` and `Permission` entities to support authorisation checks